Added Exracted DialogWindowMamanger
