+++ Assets/Dialogue/Test 1.asset
@@ -13,8 +13,8 @@ MonoBehaviour:
   m_Name: Test 1
   m_EditorClassIdentifier: 
   line:
-    speaker: {fileID: 0}
-    content: 
+    speaker: {fileID: 11400000, guid: a1327decf85425b4e909f4c3385c5334, type: 2}
+    content: dsadsa
   display: Test Choice
   response: {fileID: 11400000, guid: 4e1b5511d3201c848a48121d19a8ddb7, type: 2}
   callback:
