+++ Assets/Scripts/Dialog/Button.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: ec59b335f201ab44cbc48dac767cb0ff
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
