+++ Assets/Scripts/Dialog/ChoiceEventMapper.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 4cfdaffab1a64354fa4a0fc2b96c3ebd
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
