+++ Assets/Scripts/Dialog/ChoicesController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d2fdabba2a25b9d45bdd904edb9e0985
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
