+++ Assets/Scripts/Dialog/DialogWindowManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f69b5c71e7462f14b8a5f604a8b13959
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
