+++ Assets/Scripts/Dialog/DialogueEditor.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8659b20a6c4639141866b11f8208306f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
