+++ Assets/Scripts/Dialog/IDialogWindowManager.cs
@@ -1,0 +1,9 @@
+namespace Dialog
+{
+    public interface IDialogWindowManager
+    {
+        void ToggleDialogWindow(bool visible);
+        void ShowDialogBox();
+        void ShowChoicesBox();
+    }
+}
