+++ Assets/Scripts/Dialog/IDialogWindowManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 748cf90ab322a1b469db3e3c1aecaf80
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
