--- Assets/Scripts/Dialogue.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 6f049c50fcd3fa6468a4abaee9316c23
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
