--- Assets/Scripts/Dialogue/Actor.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 9e54eb3c06762a44ea773d68b7e0bdb9
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
