--- Assets/Scripts/Dialogue/Button.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: ec59b335f201ab44cbc48dac767cb0ff
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
