--- Assets/Scripts/Dialogue/Choice.cs
@@ -1,15 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-using UnityEngine.Events;
-
-[CreateAssetMenu(fileName = "Data", menuName = "Dialog/Choice")]
-public class Choice : ScriptableObject
-{
-    public Line line;
-    public string display;
-    public Dialogue response;
-
-    [HideInInspector]
-    public UnityEvent callback;
-}
