--- Assets/Scripts/Dialogue/Choice.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 4a788ab7df4647046a63ac76b7c62a2a
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
