--- Assets/Scripts/Dialogue/Dialog.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 61635467ad5b6be469b9daf7281d0472
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
