--- Assets/Scripts/Dialogue/Dialogue.asmdef.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 02711c3755c1a8e42abd3ca6dd221128
-AssemblyDefinitionImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
