--- Assets/Scripts/Dialogue/Dialogue.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: d7917b69129d07b4f8f418f5a39c154e
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
