--- Assets/Scripts/Dialogue/DialogueController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: c2305ca43550fe94bb44c9c38f25194c
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
